Add an rc.d script for the user-mode.
This script is based on the rc.d script delivered with the package
emulators/qemu-user-static. The script was reduced to riscv64 and
riscv64c binaries. The riscv64 magic and mask were extended to include
Elf64_Ehdr.e_flags to check if a binary includes the EF_RISCV_CHERIABI
flag and hence is a riscv64c binary.

This change allows us to run plain RISC-V and CheriABI RISC-V binaries
on the same host.