Copy files instead of creating symlinks.
The repository contains port options that are mounted by
poudriere-bulk(8) in /var/db/ports. Such files cannot be used for
symlinks since the files the symlinks would point to don't exist in
Poudriere jails.

We could make lists of files and directories that could be used for
symlinks and must be copied. Considering that Poudriere hosts are
dedicated for package building, such change would overcomplicate
bootstrapping process.

Instead, let's copy all files from the repository to a host.