Add overlay for an amd64 jail
The amd64 jail can be used to build packages with the local base set to
/usr/local64 that would match hybrid ABI packages for CheriBSD.
Such packages can be used as replacement packages to build CheriABI
packages using the QEMU user mode for CheriABI on an amd64 host.

The toolchain name indicates that it's for aarch64 but it works for any
native ABI build. This should be cleaned up in a separate commit.