Use bzip2 to build devel/gettext-tools for purecap
devel/gettext-tools must be configured not to use xz as well as git to
create an archive that is included in its package. When both are
disabled, the build system uses bzip2 instead.

Additionally, the port must be instructed to use a bzip2-specific
archive extension to correctly generate a pkg-plist file.